removed warning about signed/unsigned compare
svn-id: https://kwave.svn.sourceforge.net/svnroot/kwave/trunk@1106 bbf32e5a-961d-0410-9e10-91f32741c31e

+++ libgui/HMSTimeWidget.cpp
@@ -44,7 +44,7 @@ int HMSTimeWidget::value()
 void HMSTimeWidget::setValue(int value)
 {
     if (value < 0) value = 0;
-    if (value > m_limit) value = m_limit;
+    if ((unsigned int)value > m_limit) value = m_limit;
     m_time = value;
 
     const int seconds = (m_time % 60);
